+++ tests/unit/test_deprecations.py
@@ -16,11 +16,6 @@ class TestDeprecation(UnitTest):
         with pytest.deprecated_call():
             reddit.subreddit("all").modmail.conversations(after="after")
 
-    def test_gild_method(self, reddit):
-        with pytest.raises(DeprecationWarning) as excinfo:
-            reddit.submission("1234").gild()
-            assert excinfo.value.args[0] == "'.gild' has been renamed to '.award'."
-
     def test_reddit_token_manager(self):
         with pytest.raises(DeprecationWarning):
             Reddit(