+++ CHANGES.rst
@@ -9,6 +9,7 @@ Unreleased
 **Removed**
 
 - Remove ``APIException`` class.
+- Remove ``Comment.gild`` and ``Submission.gild`` methods.
 - Remove ``Subreddits.gold`` method.
 - Remove ``Reddit.validate_on_submit`` configuration attribute.
 - Remove ``WebSocketException.original_exception`` method.
@@ -307,8 +308,8 @@ Unreleased
 **Deprecated**
 
 - ``Subreddits.gold`` is superseded by :meth:`.Subreddits.premium`.
-- :meth:`.Submission.gild` is superseded by :meth:`.Submission.award`.
-- :meth:`.Comment.gild` is superseded by :meth:`.Comment.award`.
+- ``.Submission.gild`` is superseded by :meth:`.Submission.award`.
+- ``.Comment.gild`` is superseded by :meth:`.Comment.award`.
 
 **Fixed**